% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df2geneset}
\alias{df2geneset}
\title{Construction of gene sets}
\usage{
df2geneset(df, confidence_levels = c("ABCDE"), organism = "hs")
}
\arguments{
\item{df}{A table from DoRothEA package.}

\item{confidence_levels}{Combination of DoRothEA's confidence levels.
Corresponding regulons/interactions will be returned. This combination
can contain any combination from the letters A-E and must be encoded as a
single character string.}

\item{organism}{A character string indicating whether human DoRothEA (hs) or
mouse DoRothEA (mm) is desired.}
}
\value{
Regulons in GSEA format (see \code{\link[=fgsea]{fgsea::fgsea()}}).
}
\description{
This function coverts DoRothEA's regulons that are stored in a table to the
format required for GSEA methods (e.g. \code{\link[=fgsea]{fgsea::fgsea()}}).
}
