% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dorothea-package.R
\docType{package}
\name{dorothea-package}
\alias{dorothea}
\alias{dorothea-package}
\title{dorothea: Collection Of Human And Mouse TF Regulons}
\description{
{
\if{html}{\figure{tool_logo.png}{options: align='right' alt='logo' width='120'}}
This package contains human and mouse TF regulons.

}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://saezlab.github.io/dorothea/}
  \item \url{https://github.com/saezlab/dorothea}
  \item Report bugs at \url{https://github.com/saezlab/dorothea/issues}
}

}
\author{
\strong{Maintainer}: Pau Badia-i-Mompel \email{pau.badia@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-1004-3923}{ORCID})

Authors:
\itemize{
  \item Daniel Dimitrov \email{daniel.dimitrov@uni-heidelberg.de} (\href{https://orcid.org/0000-0002-5197-2112}{ORCID})
  \item Christian H. Holland \email{cholland2408@gmail.com} (\href{https://orcid.org/0000-0002-3060-5786}{ORCID})
  \item Luz Garcia-Alonso
}

Other contributors:
\itemize{
  \item Alberto Valdeolivas [contributor]
  \item Minoo Ashtiani [contributor]
  \item Attila Gabor [contributor]
}

}
\keyword{internal}
