% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dorothea_hs}
\alias{dorothea_hs}
\title{Human DoRothEA}
\format{A table of human TF-target interactions:
\describe{
    \item{tf}{TF identifier as HGNC symbols}
    \item{confidence}{Confidence score classifying regulons based on their quality}
    \item{target}{target identifier as HGNC symbols}
    \item{mor}{mode of regulation indicating the effect of a TF on the target}
}}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6673718/}
}
\description{
A table reporting signed human TF-target interactions. This database covers
  in total 1395 TFs targeting 20,244 genes with 486,676 unique interactions.
  In addition, each TF is accompanied with an emperical confidence level that
  was derived from the number of supporting evidences for this
  TF/interaction. The range is from A (high quality) to E (low quality).
}
\keyword{datasets}
