% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dorothea_hs_pancancer}
\alias{dorothea_hs_pancancer}
\title{Human DoRothEA (pancancer)}
\format{
A table of human TF-target interactions:
\describe{
    \item{tf}{TF identifier as HGNC symbols}
    \item{confidence}{Summary confidence score classifying regulons based on
        their quality}
    \item{target}{target identifier as HGNC symbols}
    \item{mor}{mode of regulation indicating the effect of a TF on the
        target}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6673718/}
}
\description{
A table reporting signed human TF-target interactions for cancer application. 
  The difference to "dorothea_hs" is that TGCA gene expression data was used 
  instead of GTEx to infer the networks with ARACNE. This database covers
  in total 1344 TFs targeting 20,582 genes with 213,230 unique interactions.
  In addition, each TF is accompanied with an empirical confidence level that
  was derived from the number of supporting evidences for this
  TF/interaction. The range is from A (high quality) to E (low quality).
}
\keyword{datasets}
