% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{dorothea_mm_pancancer}
\alias{dorothea_mm_pancancer}
\title{Mouse DoRothEA (pancancer)}
\format{
A table of mouse TF-target interactions:
\describe{
    \item{tf}{TF identifier as MGI symbols}
    \item{confidence}{summary confidence score classifying regulons based on
        their quality}
    \item{target}{target identifier as MGI symbols}
    \item{mor}{mode of regulation indicating the effect of a TF on the
        target}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/pubmed/31525460}
}
\description{
A table reporting signed human TF-target interactions. The difference to 
  "dorothea_mm" is that TGCA gene expression data was used instead of GTEx to 
  infer the networks with ARACNE.This database covers in total 1096 TFs 
  targeting 17,695 genes with 187,955 unique interactions. In addition, each 
  TF is accompanied with an empirical confidence level that was derived from 
  the number of supporting evidences for this TF/interaction. The range is 
  from A (high quality) to E (low quality).
}
\keyword{datasets}
