% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{entire_database}
\alias{entire_database}
\title{Entire database with associated meta data}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC6673718/}
}
\description{
This table lists all human TF-target interactions that were derived from the
  four lines of evidences. Each interaction is assigned a confidence score
  based on the number of supporting evidences. The table provides also all
  required information to trace back the origin of the interaction.
}
\keyword{datasets}
