% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{parse_confidence_levels}
\alias{parse_confidence_levels}
\title{Parsing confidende levels}
\usage{
parse_confidence_levels(confidence_levels)
}
\arguments{
\item{confidence_levels}{Combination of DoRothEA's confidence levels. This
combination can contain any combination from the letters A-E and must be
encoded as a single character string.}
}
\value{
Vector of confidence levels.
}
\description{
Parsing confidende levels
}
