% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{scViper}
\alias{scViper}
\title{VIPER wrapper for single cell data}
\usage{
scViper(
  InputObject,
  regulon,
  assay_name = "RNA",
  return_assay = FALSE,
  options = list()
)
}
\arguments{
\item{InputObject}{An expression matrix with genes (HGNC/MGI symbols) in rows
and cells in columns. It also accepts \code{Seurat} object from which the
normalised expresion matrix of the selected assay will be extracted}

\item{regulon}{Object of class regulon. Check viper documentation for
further information.}

\item{assay_name}{it only applies if the input is a Seurat object. It selects
the name of the assay on which Viper will be run. Default to:
RNA, i.e. normalized expression values.}

\item{return_assay}{it only applies if the input is a Seurat object. A
logical value indicating whether to return Dorothea results as a new
assay called Dorothea in the Seurat object used as input.
Default to FALSE.}

\item{options}{A list of named options to pass to
\code{\link[=viper]{viper::viper()}} such as \code{minsize} or
\code{method}. These options should not include \code{eset} or
\code{regulon}.}
}
\value{
A matrix containing the activity of the different TFs provided in
the regulon object.
}
\description{
This function is a convenient wrapper to run the
\code{\link[=viper]{viper::viper()}} function to apply it on single cell
data combined with Dorothea regulons
}
