% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{df2regulon}
\alias{df2regulon}
\title{Construction of dorothea regulons for viper analysis}
\usage{
df2regulon(df)
}
\arguments{
\item{df}{A regulon table from dorothea package.}
}
\value{
Regulons in the \code{viper} format.
}
\description{
This function converts DoRothEA's regulons that are stored in a table to the
format required by the \code{\link[viper]{viper}} function.
}
\examples{
# acessing (human) dorothea regulons
# for mouse regulons: data(dorothea_mm, package = "dorothea")
data(dorothea_hs, package = "dorothea")
# convert to the format required by viper
viper_regulons = df2regulon(dorothea_hs)
}
